LUCENE-5100: BaseDocIdSetTestCase.
git-svn-id: https://svn.apache.org/repos/asf/lucene/dev/trunk@1502448 13f79535-47bb-0310-9956-ffa450edef68

+++ lucene/core/src/java/org/apache/lucene/util/packed/EliasFanoDocIdSet.java
@@ -33,7 +33,7 @@ public class EliasFanoDocIdSet extends DocIdSet {
    * @param numValues The number of values that can be encoded.
    * @param upperBound  At least the highest value that will be encoded.
    */
-  public EliasFanoDocIdSet(long numValues, long upperBound) {
+  public EliasFanoDocIdSet(int numValues, int upperBound) {
     efEncoder = new EliasFanoEncoder(numValues, upperBound);
   }
